+++ frontend/src/components/User/LoginForm.js
@@ -14,8 +14,8 @@ import Switch from '@material-ui/core/Switch';
 import Button from '@material-ui/core/Button';
 import Visibility from '@material-ui/icons/Visibility';
 import VisibilityOff from '@material-ui/icons/VisibilityOff';
-import Link from './Link';
-import redirect from '../utils/redirect';
+import Link from '../Link';
+import redirect from '../../utils/redirect';
 
 const styles = theme => ({
   loginBox: {
